remove RSCRoots and replace its test with registerServerComponent tests
